refactor: reorganize api/ subdirectories (step 4.5-4.14)
Move domain-specific files from bindings/ to api/ subdirectories:
- api/s3/: S3 client and error handling
- api/secrets/: Platform secrets management
- api/ffi/: Foreign Function Interface
- api/sqlite/: SQLite bindings (moved from bindings/sqlite/)
- api/sql/: SQL client
- api/shell/: Shell bindings
- api/ipc/: IPC bindings
- api/test/: Testing utilities
- api/server/: HTTP server bindings
- api/plugin/: Bundler plugin system

Part of source code reorganization to make code location "guessable".
