+++ src/buntime/api/s3/S3Stat.classes.ts
@@ -1,4 +1,4 @@
-import { define } from "../../codegen/class-definitions";
+import { define } from "../../../codegen/class-definitions";
 
 export default [
   define({
