+++ src/buntime/bindings/JSValue.zig
@@ -2399,7 +2399,7 @@ extern "c" fn Bun__JSValue__isAsyncContextFrame(value: JSValue) bool;
 
 const string = []const u8;
 
-const FFI = @import("./FFI.zig");
+const FFI = @import("../api/ffi/FFI.zig");
 const std = @import("std");
 const JestPrettyFormat = @import("../../test_runner/pretty_format.zig").JestPrettyFormat;
